Introduce gettext (i18n library): en, ru, and sv
We want to support a few languages in the UI, so we need to leverage
abilities of wonderfull Gettext package.

We used `gettext/1` in a few places in our templates, and generate three
locale files. In addition, we also set default locale in the
configuration.

In future, we are looking into some sort of switch between available
locales.