Introduce errors handling for search requests
We want to show users feedback if their search fails for some reason
(API is not available, or word they're looking for is not in the
dictionary).

Here we introduced structured responses in our Lexin
Client/Parser/Service chain of calls. Plus, we handle possible error
cases in the DictionaryLive view by showing them as flash alerts.