Update "listen" link, provide href
Users want to be able to download mp3-files with the pronounciation (or
send links to these files), so we decided to put audio file url into
href attribute. By click, it still just plays the sound, but user now
can click right button and select "Open in a new tab", for example, or
"Copy link".