Refactor components a bit, move form to its own
We are trying to use LiveView components more: one of the first steps is
to extract search form into its own LiveComponent that handles events by
itself.

We do not need to introduce a new mechanism of broadcasting events to
parent live view, because search form updates GET-params in which are
getting handled by DictionaryLive afterwards.