Prepare app to be dockerized
Add 2-staged Dockerfile (build and app) that prepares a release of the
app and makes it self-executable to simplify running it via Docker.

We also had to update mix.exs and remove redundant release options that
we do not need now. With this approach we need to include ERTS and don't
need to generate tar.gz-file of release – it will be possible to
distribute Docker-image file (`docker save lexin > lexin.tar`) to remote
machines, load it there (`docker load < lexin.tar`) and run it.