Add PoC for SQLite files generator
Lookup of the words in the raw XML files takes a lot of time (~5 seconds
per 18MB file), so we are searching for a quicker approach.

One of the ideas is to leverage SQLite because it's simple, and quite
fast at the same time. We also need to use it only for reading – we
won't write anything back when Lexin app is running.

It's a quite clear cache for fast word lookups.

In this commit, we put some code for converting given Lexin's XML files
into SQLite database files.

So far, it works, and lookups take 10-12ms in the SQLite format.