Update development Docker image (add build-base)
In order to support our development and build environment in Docker,
we need to make it possible to compile `exqlite` package, which uses
Erlang NIFs and has to be compiled using `make` and `gcc`.

So, we add `build-base` package to our Docker dev and build images
for the compilation steps to pass.

Unfortunately, it increases dev image size from 70MB to the whopping
320MB :(