Add `word` field to the `definitions` table
Previously, having `word` field only in the `vocabulary` table allowed
us to generate suggestions only for Swedish words, because `vocabulary`
is filled only with variants of Swedish words from original
dictionaries.

In future, we want to provide an ability as in original Lexin service –
users can type in in both languages (Swedish and target one) and get
available suggestions in place to quickly find words they want.

With that said, we extract translation of the Swedish word and store it
separately – just to simplify access to it in our future steps.