Replace Lexin API client with local Dictionary
With these changes we are removing previous online Lexin API client code
with our local Dictionary service that relies on the SQLite files data
and stored XML-snippets of words' definitions inside.

We cleaned up old Tesla modules, and with a few changes to modules names
and structure now we use `Lexin.Dictionary.lookup/2` to get parsed and
structured definitions.