Update UI and translations, add more languages
One of the last steps in our migration to SQLite data-files is to start
using them in the definition card component.

In addition, we also found a few more languages that we can convert to
SQLite databases and use them.

Also, in the XML files we do not have full URLs to the services where
audio- and picture-files lie, so we have to add some wrapping code to
handle links correctly.