Fix #2: Put most relevant definition on top
Our users noticed the difference between original Lexin service output
and ours – in Lexin Light the most relevant definition was not on the
firts place in the results.

This is due to how we select data from the dictionary database files –
we do not apply any particular order, and it selects just by the way how
definitions are stored there (basically, just alphabetically).

With Swedish way of making compounds and current dictionary database
structure, it's a little bit tricky to put the most relevant definition
simply in SQL-query.

So, in this attempt we are trying to pick the most relevant (spelled
exactly as user's query – with some sanitizing) and put it first in the
results list.

We tested it with "bil", "fordon", "dammsugare", and "motor". Looks like
it work fine. But we might change it to SQL in future – we left TODO.