Add Graminfo field to be parsed and shown
We noticed a lack of grammar information in Lexin Light that is
presented in the original service.

There is optional field `<Graminfo>` in the `<Baselang>` entities in the
original XML dictionaries, which we haven't noticed previously.

We added it to the types in our app, and to the `Parser` code. We also
render this information in the search results page (with a little
re-formatting of original string – as we understood it observing XML
files data and output of original Lexin app).

Fixes #4.