Fix antonyms parser and show synonyms correctly
We wrongly parsed `<Antonym>` in the original XML – we had to take its
`Value` attribute value, not the content of the tag. Now it is fixed.

In addition, we also moved rendering of synonyms out of another `if`
statement.

P.S. Plus, we have to revisit our parser later and adjust its logic
according to the LexinSchema.xsd file, as we noticed some discrepancies
from our current understanding of its internal schema.

P.P.S. Fixes #3.