Assign previously saved lang in the `mount` callback
We want to initialize DictionaryLive component state at the `mount`
stage, and try to avoid redundant `mount`-hook in JavaScript to simplify
state-management from LiveView.

So, we refactored the initialization, send previously saved language
value at the moment of LiveView socket connection – and handle it on the
Elixir side.

The idea came after reading this post:
https://fly.io/phoenix-files/restore-liveview-state-on-startup/