Normalize search query (downcase)
Users noticed that search is case-sensitive in some cases: while it
works fine for English-like words, the case of the query affects queries
in Russian. For example, "человек" gave the results, while "Человек" was
empty.

Maybe SQLite is case-sensitive for non-ASCII characters.

We decided to downcase the search query before looking for definitions.