Add `usage` to language block, parse it and show
Users noticed lack of some more information in the Lexin Light output
(in comparison to original Lexin service). It was `<Usage/>` from the
XML files.

Here, we add this field to the parser and show it to the end users.
There is a new translation added for new label in the output.

P.S. We added a mini-task to the `mix.exs` just to remember how to update
all gettext files with one command (it parses all source code and does
corresponding updates to the translation files – see in the commit).

P.P.S. Fixes #13.