Introduce basic search suggestions
We want to add search suggestions (similar to autocompletion, but not
works as a dropdown with the words user might be typing in).

In this commit we have implemented a basic way of getting and showing
search suggestions to the user.

We still need to polish it (like to debounce), and test in different
browsers.