Fix issue no suggestions after click logo
After tapping or clicking the logo we fully refresh the page – but for
some reason our frontend sends the "blur" (maybe there is a race
condition between loosing the focus from the field and re-mounting live
view).

Here, we explicitly set `in_focus` to `true` when user is changing the
form (typing in the query, for example).