Introduce Wallaby and integration tests
As we wanted to cover Lexin user interface with some kind of tests, we
chose to write integration ones with the help of Wallaby package.

With these changes we introduce the Wallaby package, its basic
configuration and our first integration test suite SearchTest that
covers basic interactive functionality.

We have also introduced new language to the test fixtures – in order to
cover switching between the languages. Plus, we added a couple more
words into Russian fixture dictionary: to check that users can see all
related words definitions and to sync it with English XML/SQLite.