Introduce Wallaby and integration tests (#19)
As we wanted to cover Lexin user interface with some kind of tests,
we chose to write integration ones with the help of Wallaby package.

With these changes we introduced the that package, its basic
configuration and our first integration test suite `SearchTest` that
covers basic interactive functionality.

We have also introduced new language to the test fixtures – in order
to cover switching between the languages. Plus, we added a couple
more words into Russian fixture dictionary: to check that users can
see all related words definitions and to sync it with English XML/SQLite.

### Extra

We have also bumped Elixir/Erlang/Node.js/Alphine versions to the
latest. As well, as bumped all dependencies versions.

We had to introduce some refactoring and fixes to the
`SearchFormComponent`, as it started to glitch after updates of
`phoenix_live_view package`.

Added section about Wallaby and used `vale` to clean up syntax and
re-write a few sentences.