Fix search form re-connection issues
We noticed that when reconnection happens user gets into endless
loading loop. This can happen, for example, when a laptop wents to
sleep, you wake it up, and the websocket on the Lexin page reconnects to
the server.

In the logs we noticed that front-end was always trying to send
`suggest` event with only `lang` (and `_target` as meta) value. We do
not have `handle_event` for this case, and our backend always got
crashed.

At first, we didn't get _why_ front-end even sends this `suggest` event.

After investigation, we found that it's a normal LiveView behaviour and
it attempts to send `phx-change` event in case of crash or reconnection,
read more: https://hexdocs.pm/phoenix_live_view/form-bindings.html#recovery-following-crashes-or-disconnects

We fixed it by moving `phx-change` attribue exactly to the `input` field
from the `form` itself. Which is logically correct for us – we want to
show suggestions only when user changes the query.

P.S. We have also returned `phx-debounce` setting for the suggestion
delay. It was accidentally removed recently.

P.P.S. Though, as we also noticed that with the `phx-change` event
re-conciliation we were receiving not all the fields of the form, but
only `select` one (which does have its own `phx-change` handler
configured). This might be the bug, but it is going to be handled
separately.