Introduce GitHub Actions workflow for testing
In order to leverage power of CI/CD (and free minutes that GitHub
generously provides to public repositories) as well as in strenghtening
the development process, we are introducing GitHub Actions workflow.

It is supposed to automate test runs on every Pull Request (to master),
as well as being initiated after each push to master.