Replace a tag to the .link component (and patch)
The top left logo on the search form page (the home page) behaved a bit
weird before this changes: when user clicks on the link, browser does a
full refresh of the page, including reconnection to the websocket and
re-mounting general live view (the search form); this led to the
noticeable "jumpy" change in the language selector (because we pre-fill
it with the lang stored in user's localStorage).

We have used a standard .link live view component and instructed it to do
the patch action (see docs), so we don't re-mount live view and there is
no "jumpy" behavior of the lang selector anymore.