Replace live CardComponentn with static Phoenix.Component one
To allow refreshing page when user changes languages back to previously
selected (in frames of the same session; without page refresh), we
replaced old live CardComponent with SerpComponents.definition/1 static
component.

It was a mistake to make these definition cards as live components in
the first place, as we never used any of "live" functionality and used
them just to place partial templates.