Add some notes on References (and their Types)
Previously, we looked over some XML tags that are optional (but still
used from time to time, and as we found out the information these tags
contain is quite useful).

We missed `<Reference TYPE="..." VALUE="..."/>` tags and do not parse
them, so we do not show information from them either.

We started to research this thing and wrote a few notes on our findings.
Hopefully, this could be useful in future.