Replace listen link (anchor tag) with button
We noticed a bug: when user clicks the old listen link (made with a-tag
and "onclick" attribute that runs a snippet of JS (that returns `false`
in the end of execution), our form stops working (no submit, not reset
works) until user clicks on the logo and reload the main page.

By replacing that link with a `button` tag component, we solved the
issue.

In addition, we have also styled it with a "speaker" icon and slightly
distinguishable UI look from other links.