Use <script> to preselect saved user language
To avoid (a bit) annoying language dropdown UI jumps when page is
loading (before we set it through LiveView) we decided to write a tiny
JS hack that loads previously saved language from local storage and set
it as soon as possible in the language selector.

For new users no language will be set and they still see "Valj sprak"
option in this dropdown.