Improve suggestions list (quality and order)
Previously, we didn't pay too much attention to the suggestions list
that we get from SQLite and for the request like "bil" we would show
users quite unrelevant suggestions at the top of the list (especially,
due to the fact that we limit this list length). As for "bil", we were
showing these suggestions:

- bildad
- bildade
- bildat
- bildning
- bildningen

With these changes, we now get all the similar words from SQLite (that
start as our query) and then sort them by similarity (Elixir has a great
helper for that job – `String.bag_distance/2`). So, now for the same
"bil" query, users will see these suggestions:

- bil
- bila
- bild
- bill
- bilen
- bilar
- bilan
- bilor

P.S. We have also increased the default number from 5 to 8 items in the
list to show to a user.