Add 'preferred lang' feature to pick the lang
As we found out, some users do not notice out language selector
dropdown, and they stuck when they see 'Sprak stods inte' flash message
while search for a word.

We want to avoid cases like that and here is what we came up with:

- we check browser's language, check if any language was already saved
  during previous sessions, check if a user opened someone's else result
  page (that sets the lang parameter in URL)
- priority to set the preferred language is the next (if any is set):
  1. from URL
  2. previously saved in local storage
  3. browser's preferred
  4. default - 'english'
- we also removed 'Select your language' option at all from the
  dropdown, so we won't show the alert 'Sprak stods inte' – there is
  always must be at least some language selected

A note regarding the last point: it is possible to visit a URL with
mistyped language parameter value (the one that we do not support) – in
cases like this, users will see 'Sprak stod inte' alert.

fixup: JS tiny bits