Fix ArgumentError `:unicode.characters_to_nfc_list/1`
The error caught by Sentry. There were also a few additional errors
that Sentry caught. Original issue URL:

https://lexin-mobi.sentry.io/share/issue/1dbb6b0a69e346e6b627f009a24457c7/

A little digging in the area revealed that in some rare cases (e.g., for
Swedish `stavar`), the `<phonetic>...</phonetic>` block of the
definition doesn't have the `file` attribute set.

This led to the issue where code was trying to convert the input string,
but getting `nil` instead.

To fix that, we updated `Parser`, `SerpComponents`, and `Phonetic`
modules. From now we show `listen` button optionally: only if the file
URL exists and not nil.