Add SitemapTester module and mix task
To check generated URLs we have made this quite straightforward, dumb,
but useful crawler that checks availability of all the URLs we listed in
the newly generated sitemaps.

During initial testings on localhost, we also noticed some weird issues
with words that contain "/" character (`c/o`, `CAD/CAD`, and couple of
other). So, we fixed it in the sitemap generator code.