Fix failures on reference-only word definitions
During sitemap testing, we noticed that around 10 words make our app
fail.

The words are these: akne, aknen, idag, maskara, satin, schyst, shejk,
shejken, shejker, silhuett.

What is common about them? One (or single) of their definitions contains
a reference-only link to the other definition (of a similarly written
word).

Our definition card was not ready to render data like this, so we had to
add a few guards in the template.

In addition, we have also adjusted SEO-title generation, as in some
dictionaries, definitions do not have a translation, so we were getting
"broken" page titles.

We have updated our test fixtures with production data of one of such a
word – silhuett - and covered title and rendering cases with tests.