Use linux/amd64 platform for the release images...
...to avoid issues with the default arm64 that is being used when
building the image on M-based macs.

To run it smoothly we had to use `ERL_FLAGS="+JMsingle true" workaround
(due to the issue with Erlang's JIT running inside a QEMU, which must be
used on M-based macs to build an amd64 image).

In addition, we also slightly cleaned the variables in the `build.sh`
file.