Refactor sitemap files generation and handling
We thought that it is too much to add static sitemap files into the
final Docker images with releases. Especially, that these files don't
need to be re-generated between every single release: because
dictionaries stay the same.

However, it would be better if we still serve the files from the `/`
path of the website, so we shall try that new configuration with
`PlugStatic` and external directory.

External directory in its turn is needed for running images in Docker –
we can mount an external directory with sitemaps in the similar way as
we do for `DICTIONARIES_ROOT`.

In addition to the main changes, we have also done various clean ups in
the related pieces of code, little performance improvements to sitemap
generation process, etc.