Update way of access to Dictionaries and Sitemaps...
...we stuck into issue with the Plug.Static configuration – it cannot be
configured at a runtime (using environment variables which are picked in
the `runtime.exs`). At least, not without a customization (which we
would like to avoid).

The `:from` option is being set at the compilation step, so it's `nil`
when we run the app via Docker container.

So, we thought that we can use relative paths to dictionaries (and
sitemaps) and re-wrote all the configs, moved local files into new
places, and did other corresponding changes.