Re-write Lexin.Dictionary and its descendants
In the previous implmentation, we had one GenServer whose state
maintained ALL the opened SQLite connections. This was error-prone: if
one SQL request failed to run on the database, it lead to the dead of
the whole process and other dictionarie's connections.

In this new version, we are adopting Registry and Supervisors to build a
controllable tree of (searchable) worker processes. Each of these
workers maintains a single dictionary database connection (by language)
and provides access to the data. If one dies (for some reason),
Repo.Supervisor will try to get it back up and running, so only a single
(or little amount) of search requests shall be affected.