Add OpenGraph images cache (in a directory)
Observations during the development shows that generation of a new PNG
image (from the given SVG template) takes ~180-200ms in average (on M3).

That's quite a significant amount of time, especially, if some bots
start to do too many requests (to the same image). So, we decided that
we can add a new directory where we will put generated images and serve
them from there if they exist, or generate and save new ones.

P.S. This approach won't help if someone decides to request different
words.

P.P.S. We would like to monitor how it goes in production, but we would
need to add some auto-clean job that will periodically check size of the
cache directory - to prevent its uncontrolled growth.