Replace previous naive SEO code with less naive...
...phoenix_seo library and approach based on Elixir protocols.

We had to rewrite LexinWeb.SEO module from scratch, update layout, a do
other small updates here and there.

In addition to the new library for the Search Engine Optimizations, we
are trying to change the page titles, descriptions, and other relevant
meta information (simplify it, make it more useful for end users).

We have also added Inkscape version for the default OpenGraph image, as
well as its (lossless) optimized version of the PNG exported from the
Inkscape as default OpenGraph image to be used as fallback (if it has
not been set explicitly on a page).