Replace "Lexin Mobi" with "Lexin.mobi"
With the recent OpenGraph image updates (and logo updates that happened
the same time), we went with the new name for the website, we started to
use a dot in the middle and write "mobi" downcased.

We decided to update other places that are still use old name.

In addition, we have also updated logo and Apple icons. The icons also
got updates recently, but we missed the white background for them.