Fix flaky integration tests with (dirty) workaround
Lately, we have noticed some level of flakyness of the integration tests
we run using Wallaby and Chromedriver. They all behave like a race
condition in updating HTML titles of the page via LiveView events.

A few links here https://github.com/elixir-wallaby/wallaby/issues/804
led to assumption that it might be related to the Chromedriver itself,
but we cannot dig deeper now.

We decided to go with a dirty workaround using Process.sleep(50)
(50 milliseconds) that lets to wait the controlling process before it
asserts the expected changes to the page title. Multiple local runs (in
the same environment that was showing flaky behavior just before the
changes) show that it works.