Introduce general (and basic) rate limiter...
...to the whole endpoint.

In August 2025, we started to notice something like DDoS attacks, or
maybe just unreasonable bots tried to scrape the dictionaries content
from us.

This led to errors due to timeouts (we were able to reproduce the issue
only by running `siege` with at least 100 concurrent connections). Thus
we overused Sentry's quota.

The introduction of the rate limiter is the first step in solving the
current situation.