Use remote_ip plug to fix rate limiter
We've noticed a weird behavior of the configured rate limiter; it is
prohibiting even rare requests from some addresses. Some debugging
revealed that in the `Lexin.RateLimiter` ETS table we had only two
records:

```
> :ets.tab2list(Lexin.RateLimit)
[
  {{"web_requests:::ffff:172.20.0.1", 29404156}, 217, 1764249420000},
  {{"web_requests:::ffff:172.20.0.1", 29404157}, 25, 1764249480000}
]
```

This helped us to realize that `:inet.ntoa(conn.remote_ip)` instruction
that we use to build the rate limiter's key takes not the external
client's IP address but an internal Docker one (because we run Lexin in
a container).

We also run Caddy as a reverse proxy to Lexin's container, and it's
supposed to pass the `X-Forwarded-For` header by default. RemoteIP
supposedly resolves the issue about `conn.remote_ip`.