Fix external pictures links prefix
Recently, we have noticed that extra links to the pictures do not work
as intended anymore: instead of a page with a picture for the given
word, they started to redirect to the main page.

We cannot say for sure (as we do not control external libraries), but we
assume that the old external pictures library got updated because links
like https://bildetema.oslomet.no/bildetema/bildetema-html5/… redirect
now to https://nybildetema.oslomet.no/en/#/eng and all the navigation is
under the hash part of the URL (it seems it was rebuilt as a single-page
application).

At the same time, the original Lexin started to use another prefix for
the pictures' links:

https://bildtema.isof.se/ instead of the old https://bildetema.oslomet.no

> Examples of fully-defined links:
>
> - (old) https://bildetema.oslomet.no/bildetema/bildetema-html5/bildetema.html?version=swedish&languages=swe,eng&page=17&subpage=2
> - (new) https://bildtema.isof.se/bildetema/bildetema-html5/bildetema.html?version=swedish&languages=swe,eng&page=17&subpage=2

However, this new website address shows SSL-related issues (the
certificate is expired), which might indicate that its web server is not
configured properly.

So, although we changed the prefix to the new external library site, we
also want to hide the “Extra” sections until we ensure that the new site
is stable.